move get_series_to_skip to dcmd
